code-assistant/requirements.txt
openai>=1.30.0
gradio>=4.0.0
python-dotenv>=1.0.0

